Make the room owner accept requesting users
In naming for JS identifiers of the RTDB data in rooms/{ownerId}/{ownRoomId}/public_info,
it uses `internalPublic` instead of `publicInfo`.

It doesn't append a duplicated uid to accepting path in starting of auto room entering.

It show the new chat page component via the routing component.

It starts to make the room owner accept requesting users on going to his/her chat room page.

In the new chat page,
it prepares both internal info and entrance info of the target room.